+++ intelafbackend/src/main/java/com/ayd2/intelafbackend/controllers/EmployeeController.java
@@ -4,13 +4,19 @@
  */
 package com.ayd2.intelafbackend.controllers;
 
+import com.ayd2.intelafbackend.dto.employee.EmployeeRequestDTO;
+import com.ayd2.intelafbackend.dto.employee.EmployeeResponseDTO;
 import com.ayd2.intelafbackend.dto.products.ProductResponseDTO;
 import com.ayd2.intelafbackend.services.EmployeeService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -28,9 +34,33 @@ public class EmployeeController {
         this.employeeService = employeeService;
     } 
     
-//    @GetMapping("/get-role/{id_user}")
-//    public ResponseEntity<List<ProductResponseDTO>> getRoleById(@PathVariable Long id_user) {
-//        return ResponseEntity.ok(employeeService.findAll());
-//    }
+    @GetMapping("/employee-by-username/{username}")
+    public ResponseEntity<EmployeeResponseDTO> findEmployeeById(@PathVariable("username") String username){
+        return  ResponseEntity.status(HttpStatus.OK).body(employeeService.findByUsername(username));
+    }
+    
+    @GetMapping("/list-all-employees")
+    public ResponseEntity<List<EmployeeResponseDTO>> listAllEmployees() {        
+        return ResponseEntity.status(HttpStatus.CREATED).body(employeeService.findAll());
+    }
+    
+    @PostMapping("/create-employee")
+    public ResponseEntity<EmployeeResponseDTO> createEmployee(@RequestBody EmployeeRequestDTO newEmployee) {
+        EmployeeResponseDTO responseDTO = employeeService.createEmployee(newEmployee);
+        
+        return ResponseEntity
+                .status(HttpStatus.CREATED)
+                .body(responseDTO);
+    }
+    
+    @PutMapping("/update-employee/{id}")
+    public ResponseEntity<EmployeeResponseDTO> updateEmployee(@PathVariable("id") String id, @RequestBody EmployeeRequestDTO updatedEmployee) {
+        EmployeeResponseDTO responseDTO = employeeService.updateEmployee(id, updatedEmployee);
+        if (responseDTO != null) {
+            return ResponseEntity.ok(responseDTO);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
     
 }
