+++ intelafbackend/src/main/java/com/ayd2/intelafbackend/services/EmployeeService.java
@@ -4,6 +4,7 @@
  */
 package com.ayd2.intelafbackend.services;
 
+import com.ayd2.intelafbackend.dto.employee.EmployeeRequestDTO;
 import com.ayd2.intelafbackend.dto.employee.EmployeeResponseDTO;
 import java.util.List;
 
@@ -14,4 +15,8 @@ import java.util.List;
 public interface EmployeeService {
     List<EmployeeResponseDTO> findAll();
     EmployeeResponseDTO findByUsername(String username);
+    
+    EmployeeResponseDTO createEmployee(EmployeeRequestDTO newEmployee);
+
+    EmployeeResponseDTO updateEmployee(String id, EmployeeRequestDTO updatedEmployee);
 }
