+++ intelafbackend/src/main/resources/application.properties
@@ -7,3 +7,8 @@ spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.hibernate.ddl-auto=none
 server.error.include-message=always
+
+
+spring.web.cors.enabled=true
+spring.web.cors.allowed-origins=http://localhost:3000  # Replace with your frontend origin
+spring.web.cors.allowed-methods=GET,POST,PUT,DELETE,OPTIONS  # Optional: Allowed HTTP methods
